+++ CHANGELOG.md
@@ -37,6 +37,7 @@ See the [v0.3.0 milestone] for a full list of all changes.
 [#252]: https://github.com/rust-marker/marker/pull/252
 [#256]: https://github.com/rust-marker/marker/pull/256
 [#259]: https://github.com/rust-marker/marker/pull/259
+[#260]: https://github.com/rust-marker/marker/pull/260
 [#263]: https://github.com/rust-marker/marker/pull/263
 [#265]: https://github.com/rust-marker/marker/pull/265
 
@@ -53,6 +54,7 @@ See the [v0.3.0 milestone] for a full list of all changes.
 - [#244]: `StmtKind` and `PatKind` no longer wrap `Kind*` directly
 - [#245]: `emit_lint()` takes less arguments and returns a `DiagnosticBuilder` instance
 - [#263]: Updated the [`ui_test`](https://crates.io/crates/ui_test) used by `marker_uitest` from `v0.11.7` to `v0.21.2`
+- [#260]: Moved `AstContext::{body, item, lint_level_at}` to the new `AstMap` struct accessible via `MarkerContext::ast()`
 - [#265]: Removed the `CallableData` trait
 
 ### Internal